requirements.txt
requests
pandas
flask
ta
python-dotenv